fix(api): serialize Decimal score/marks as JSON numbers; add test_paper_title
Pydantic v2 serialized Decimal fields as JSON strings, so the SPA received
e.g. "0.00" where it expected a number (caused `value.toFixed is not a
function` on the attempts page). Type the client-facing money/score fields as
float instead of Decimal so they serialize as numbers:
- attempt: AttemptResultOut.{raw_score,final_score,percentile},
  AttemptSummaryOut.{final_score,percentile}, TestPaperBrief.total_marks
- test_paper: TestPaperOut.{total_marks,negative_marking_factor}

Also add `test_paper_title` to AttemptSummaryOut (joined from test_papers) so
the attempt-history UI can show the paper name without a global papers endpoint.

Contract tests unaffected (they coerce with float()): full attempt + testpaper
suites green.

+++ backend/app/schemas/test_paper.py
@@ -25,8 +25,8 @@ class TestPaperOut(BaseModel):
     title: str
     scheduled_at: datetime | None = None
     duration_seconds: int
-    total_marks: Decimal
-    negative_marking_factor: Decimal
+    total_marks: float
+    negative_marking_factor: float
     shuffle_questions: bool
     shuffle_options: bool
     created_at: datetime
